Fix incorrect pruning with duplicate letters
Consider these guesses for target word WHACK:
STONY (00000)
CAREX (22000)
ABACK (00111)

The first A in aback is reported as a missing letter by Wordle because the center A is already correct.

Make pruning more conservative in cases where a letter is missing but present elsewhere in the word. Prune subtrees where the letter is not in the correct position.